requirements.txt
PySide6
psutil
Pillow
numpy==1.26.4
opencv-contrib-python==4.11.0.86
mediapipe==0.10.21
matplotlib

pycaw
comtypes
sounddevice
faster-whisper
